+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Improvement: descriptions of the features
 - Improvement: pagination on time zone settings
 - Improvement: refresh button in chat list settings
+- Fix: messages from a linked channel were mistakenly filtered by the profanity filter
 - Fix: profanity filter ignored message which starts from command
 - Fix: profanity filter ignored profanity in forwarded names
 - Fix: order of voters in voteban results was inconsistent